+++ CHANGELOG.rst
@@ -6,9 +6,11 @@ Version v0.13.0
 
 New Features
 ~~~~~~~~~~~~~
-- Population type available in NodePopulation/EdgePopulation
 - Node/edge populations are now supported in config
+- Population type available in NodePopulation/EdgePopulation
+- Population config (if given) overwrites the "components" config for that population
 - Alternate morphology directories (.h5, .asc) are now supported
+- Added different edge/node types to bluepysnap/bbp.py
 
 Improvements
 ~~~~~~~~~~~~~~